requirements.txt
flask
openai
chromadb
python-dotenv
sendgrid
